Move BrowserSwitch module to com.braintreepayments.api package.
