+++ browser-switch/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest package="com.braintreepayments.browserswitch"
+<manifest package="com.braintreepayments.api.browserswitch"
     xmlns:android="http://schemas.android.com/apk/res/android">
     <queries>
         <intent>
