+++ browser-switch/src/main/java/com/braintreepayments/api/BrowserSwitchClient.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.browserswitch;
+package com.braintreepayments.api;
 
 import android.content.Context;
 import android.content.Intent;
