+++ browser-switch/src/main/java/com/braintreepayments/api/BrowserSwitchOptions.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.browserswitch;
+package com.braintreepayments.api;
 
 import android.net.Uri;
