+++ browser-switch/src/test/java/com/braintreepayments/api/PersistentStoreUnitTest.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.browserswitch;
+package com.braintreepayments.api;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -11,7 +11,7 @@ import org.mockito.Mockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
-import static com.braintreepayments.browserswitch.PersistentStore.PREFERENCES_KEY;
+import static com.braintreepayments.api.PersistentStore.PREFERENCES_KEY;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
