+++ demo/src/androidTest/java/com/braintreepayments/api/demo/BrowserSwitchTest.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.browserswitch.demo;
+package com.braintreepayments.api.demo;
 
 import android.app.Instrumentation;
 import android.content.Context;
