+++ demo/src/main/AndroidManifest.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.braintreepayments.browserswitch.demo">
+    package="com.braintreepayments.api.demo">
 
     <application
         android:allowBackup="false"
         android:icon="@mipmap/ic_launcher"
         android:label="Browser Switch Demo"
         android:theme="@style/AppTheme">
-        <activity android:name="com.braintreepayments.browserswitch.demo.DemoActivity">
+        <activity android:name="com.braintreepayments.api.demo.DemoActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
