+++ demo/src/main/java/com/braintreepayments/api/demo/DemoActivity.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.browserswitch.demo;
+package com.braintreepayments.api.demo;
 
 import android.os.Bundle;
